requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
confusable-homoglyphs==3.2.0
crispy-bootstrap4==2025.6
distlib==0.3.5
dj-database-url==3.0.1
Django==6.0
django-crispy-forms==2.5
django-filter==25.2
django-registration-redux==2.13
djangorestframework==3.16.1
filelock==3.8.0
gunicorn==23.0.0
joblib==1.5.2
Markdown==3.10
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==2.5.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2022.6
qrcode==8.2
reportlab==4.4.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
tzdata==2025.2
virtualenv==20.16.3
whitenoise==6.11.0
